+++ src/constants.js
@@ -1,0 +1,11 @@
+export const ERR_MISSING_TWITTER_KEYS = 'Missing required Twitter access keys!'
+export const ERR_MISSING_INPUTS =
+  'Missing required inputs: action, release or repository!'
+export const ERR_TWITTER =
+  'Sending message failed due to an error from Twitter: '
+export const MSG_REPO_IS_PRIVATE =
+  'Repository is private, no message is sent to Twitter.'
+export const MSG_REPO_IS_NOT_PUBLISHED =
+  'Repository is not published so no message is sent to Twitter.'
+export const MSG_MESSAGE_LOG = 'Sending message: '
+export const RELEASE_PUBLISHED = 'published'
