+++ test/server.test.js
@@ -1,13 +1,13 @@
 'use strict'
 
 import { TwitterApi } from 'twitter-api-v2'
-import buildServer, {
+import buildServer from '../src/server.js'
+import {
   ERR_TWITTER,
-  ERR_MISSING_TWITTER_KEYS,
   ERR_MISSING_INPUTS,
   MSG_REPO_IS_PRIVATE,
   MSG_REPO_IS_NOT_PUBLISHED
-} from './server.js'
+} from '../src/constants.js'
 
 const testServer = buildServer({
   LOG_LEVEL: 'silent',
@@ -23,30 +23,6 @@ jest.mock('twitter-api-v2')
 describe('GH app', () => {
   afterEach(() => {
     jest.clearAllMocks()
-  })
-
-  it('healthcheck route', async () => {
-    const response = await testServer.inject('/healthcheck')
-
-    expect(response.statusCode).toBe(200)
-  })
-
-  it('release route, no Twitter keys', async () => {
-    const testServerNoTwtKeys = buildServer({
-      LOG_LEVEL: 'silent',
-      PRETTY_PRINT: false
-    })
-    const response = await testServerNoTwtKeys.inject({
-      method: 'POST',
-      headers: {
-        'content-type': 'application/json'
-      },
-      url: '/release',
-      body: JSON.stringify({})
-    })
-
-    expect(response.statusCode).toBe(500)
-    expect(response.body).toBe(ERR_MISSING_TWITTER_KEYS)
   })
 
   it('release route, missing action, repository and release inputs', async () => {
@@ -121,7 +97,16 @@ describe('GH app', () => {
         }
       }
     })
-    const response = await testServer.inject({
+
+    const server = buildServer({
+      LOG_LEVEL: 'silent',
+      TWITTER_API_KEY: process.env.TWITTER_API_KEY,
+      TWITTER_API_SECRET: process.env.TWITTER_API_SECRET,
+      TWITTER_ACCESS_TOKEN: process.env.TWITTER_ACCESS_TOKEN,
+      TWITTER_ACCESS_TOKEN_SECRET: process.env.TWITTER_ACCESS_TOKEN_SECRET
+    })
+
+    const response = await server.inject({
       method: 'POST',
       headers: {
         'content-type': 'application/json'
@@ -161,7 +146,15 @@ describe('GH app', () => {
         }
       }
     })
-    const response = await testServer.inject({
+    const server = buildServer({
+      LOG_LEVEL: 'silent',
+      TWITTER_API_KEY: process.env.TWITTER_API_KEY,
+      TWITTER_API_SECRET: process.env.TWITTER_API_SECRET,
+      TWITTER_ACCESS_TOKEN: process.env.TWITTER_ACCESS_TOKEN,
+      TWITTER_ACCESS_TOKEN_SECRET: process.env.TWITTER_ACCESS_TOKEN_SECRET
+    })
+
+    const response = await server.inject({
       method: 'POST',
       headers: {
         'content-type': 'application/json'
@@ -182,32 +175,5 @@ describe('GH app', () => {
 
     expect(response.statusCode).toBe(403)
     expect(response.body).toBe(`${ERR_TWITTER} ${errMessage}`)
-  })
-
-  it('release route, server failure', async () => {
-    TwitterApi.mockImplementation(() => {
-      return null
-    })
-    const response = await testServer.inject({
-      method: 'POST',
-      headers: {
-        'content-type': 'application/json'
-      },
-      url: '/release',
-      body: JSON.stringify({
-        action: 'published',
-        repository: {
-          name: 'test_repo',
-          private: false
-        },
-        release: {
-          tag_name: '1.0.0',
-          html_url: 'http://example.com'
-        }
-      })
-    })
-
-    expect(response.statusCode).toBe(500)
-    expect(response.body).toContain('Error:')
   })
 })
